feat: add complaint status management with enum and update status functionality

+++ src/complaints/complaints.service.ts
@@ -1,15 +1,15 @@
-import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, isValidObjectId } from 'mongoose';
-import { Complaints, ComplaintsSchema } from '../../models/complaints.schema';
+import { Complaints, ComplaintsSchema, ComplaintStatus } from '../../models/complaints.schema';
 import { CreateComplaintsDTO } from './dtos/create-complaints.dto';
 import { UpdateComplaintsDTO } from './dtos/update-complaints.dto';
 import { UserRole } from '../../models/user.schema';
+import { UserPayload } from 'src/forms/forms.controller';
 
 @Injectable()
 export class ComplaintsService {
 
-  //Mapa de Destino
   private readonly destinoRoleMap: { [key: string]: string } = {
     Aulas: 'professor',
     Diretores: 'admin',
@@ -18,27 +18,46 @@ export class ComplaintsService {
     Extracurriculares: 'manager',
   };
 
-  constructor(
-    @InjectModel(Complaints.name)
-    private readonly complaintsModel: Model<ComplaintsSchema>,
-  ) { }
+  constructor(@InjectModel(Complaints.name) private readonly complaintsModel: Model<ComplaintsSchema>) { }
 
-  // Determinar o destinoRole com base no tópico
   private determineDestinoRole(topico: string): string {
     return this.destinoRoleMap[topico] || 'admin';
   }
 
-  // Permissão do usuário para modificar ou deletar denúncia
-  private checkPermisson(complaint: Complaints, userRole: UserRole): void {
+  private checkActionPermission(
+    complaint: Complaints,
+    userRole: UserRole | string
+  ):
+    void {
+
     if (userRole === UserRole.ADMIN) {
       return;
     }
     if (complaint.destinoRole !== userRole) {
-      throw new UnauthorizedException('Você não tem permissão para modificar ou deletar esta denúncia. Ela está destinada ao cargo de ' + complaint.destinoRole);
+      throw new ForbiddenException(`Você não tem permissão para modificar esta denúncia. Destino: ${complaint.destinoRole}`);
     }
   }
 
-  async findAll(topico?: string, destinoRole?: string): Promise<Complaints[]> {
+  async updateStatus(
+    complaintId: string,
+    newStatus: ComplaintStatus,
+    user: UserPayload
+  ):
+    Promise<Complaints> {
+    const complaint = await this.complaintsModel.findById(complaintId).exec();
+    if (!complaint) {
+      throw new NotFoundException('Denúncia não encontrada.');
+    }
+    this.checkActionPermission(complaint, user.role);
+    const updatedComplaint = await this.complaintsModel.findByIdAndUpdate(
+      complaintId,
+      { status: newStatus },
+      { new: true }
+    ).exec();
+    return updatedComplaint!;
+  }
+
+  async findAll(topico?: string, destinoRole?: string, status?: ComplaintStatus): Promise<Complaints[]> {
     const query: any = {};
     if (topico) {
       query.topico = topico;
@@ -46,17 +65,17 @@ export class ComplaintsService {
     if (destinoRole) {
       query.destinoRole = destinoRole;
     }
+    if (status) {
+      query.status = status;
+    }
     return this.complaintsModel.find(query).sort({ createdAt: -1 }).exec();
   }
 
   async findOne(idOrSlug: string): Promise<Complaints> {
     let complaint: Complaints | null = null;
-    // Verifica se idOrSlug é um ObjectId válido
     if (isValidObjectId(idOrSlug)) {
-      // Busca pelo id
       complaint = await this.complaintsModel.findById(idOrSlug).exec();
     }
-    // Se não encontrou pelo id, busca pelo slug
     if (!complaint) {
       complaint = await this.complaintsModel.findOne({ slug: idOrSlug }).exec();
     }
@@ -78,9 +97,7 @@ export class ComplaintsService {
 
   async update(idOrSlug: string, updateDto: UpdateComplaintsDTO, userRole: UserRole): Promise<Complaints> {
     let existingComplaint: Complaints | null = null;
-    let updatedComplaint: Complaints | null = null;
 
-    // Encontrar a denúncia existente
     if (isValidObjectId(idOrSlug)) {
       existingComplaint = await this.complaintsModel.findById(idOrSlug).exec();
     }
@@ -90,39 +107,27 @@ export class ComplaintsService {
     if (!existingComplaint) {
       throw new NotFoundException(`Denúncia com id ou slug "${idOrSlug}" não encontrada`);
     }
-
-    // Verificar permissões
-    this.checkPermisson(existingComplaint, userRole);
-
-    // Atualizar a denúncia
-    updatedComplaint = await this.complaintsModel.findByIdAndUpdate((existingComplaint as any)._id, updateDto, { new: true }).exec();
-    
+    this.checkActionPermission(existingComplaint, userRole);
+    const updatedComplaint = await this.complaintsModel.findByIdAndUpdate((existingComplaint as any)._id, updateDto, { new: true }).exec();
     if (!updatedComplaint) {
-      // Precaução
       throw new NotFoundException(`Denúncia com id ou slug "${idOrSlug}" não encontrada após permissão.`);
-      }
+    }
 
     return updatedComplaint;
   }
 
   async delete(idOrSlug: string, userRole: UserRole): Promise<{ message: string }> {
     let deletedComplaint: Complaints | null = null;
-
-    // Se for ObjectId válido, tenta deletar pelo id
     if (isValidObjectId(idOrSlug)) {
       deletedComplaint = await this.complaintsModel.findById(idOrSlug).exec();
     }
-
-    // Se não encontrou para deletar pelo id, tenta pelo slug
     if (!deletedComplaint) {
       deletedComplaint = await this.complaintsModel.findOne({ slug: idOrSlug }).exec();
     }
-
     if (!deletedComplaint) {
       throw new NotFoundException(`Denúncia com id ou slug "${idOrSlug}" não encontrada`);
     }
-
-    this.checkPermisson(deletedComplaint, userRole);
+    this.checkActionPermission(deletedComplaint, userRole);
     await this.complaintsModel.deleteOne({ _id: (deletedComplaint as any)._id }).exec();
 
     return { message: `Denúncia com id ou slug "${idOrSlug}" deletada com sucesso.` };
